refactor(#97): vehicles/api — bulletproof-react segmentation
git mv:
- commitSave{,.test}.ts, waitForVehicleInList.ts → api/
- projection/fetchVehicles{,.test}.ts, fetchVehiclesAndApply{,.test}.ts → api/
- xml/buildVehiclePairXml.ts, fetchVehicleNetexXml.ts → api/
- xml/parseVehicleImportResponse{,.test}.ts → api/
- xml/Vehicle-parser{,.test}.ts, VehicleModel-parser{,.test}.ts → api/

External call site updated: VehicleCreatePage (waitForVehicleInList).
