+++ src/data/vehicles/api/Vehicle-parser.test.ts
