+++ src/data/vehicles/api/Vehicle-parser.ts
