+++ src/data/vehicles/api/VehicleModel-parser.ts
