+++ src/data/vehicles/api/buildVehiclePairXml.ts
