+++ src/data/vehicles/api/fetchVehicleNetexXml.ts
