+++ src/data/vehicles/api/fetchVehicles.test.ts
