+++ src/data/vehicles/api/fetchVehiclesAndApply.test.ts
