+++ src/data/vehicles/api/fetchVehiclesAndApply.ts
