+++ src/data/vehicles/projection/useVehicles.ts
@@ -3,7 +3,7 @@ import { useConfig } from '../../../contexts/configContext.ts';
 import { useAuth } from '../../../auth/authUtils.ts';
 import type { Order } from '../../../components/data/dataTableTypes.ts';
 import type { VehicleGQLShaped, VehicleColumnKey } from '../types/vehicleGqlShaped.ts';
-import { fetchVehiclesAndApply } from './fetchVehiclesAndApply.ts';
+import { fetchVehiclesAndApply } from '../api/fetchVehiclesAndApply.ts';
 import { compareVehicles } from '../utils/vehicleSortValue.ts';
 
 /**
