+++ src/data/vehicles/useVehicle.ts
@@ -1,8 +1,8 @@
 import { useEffect, useState, useCallback } from 'react';
 import { useConfig } from '../../contexts/configContext.ts';
 import { useAuth } from '../../auth/authUtils.ts';
-import { fetchVehicleNetexXml } from './xml/fetchVehicleNetexXml';
-import { parseVehicleXml } from './xml/Vehicle-parser';
+import { fetchVehicleNetexXml } from './api/fetchVehicleNetexXml';
+import { parseVehicleXml } from './api/Vehicle-parser';
 import type { Vehicle } from './types/Vehicle';
 
 /**
