+++ src/data/vehicles/xml/useVehiclePairSave.ts
@@ -2,8 +2,8 @@ import { useCallback, useState } from 'react';
 import { useAuth } from '../../../auth/authUtils';
 import { useConfig } from '../../../contexts/configContext';
 import { importAsNetexToBackend } from '../../vehicle-imports/vehicleImportServices';
-import { buildVehiclePairXml } from './buildVehiclePairXml';
-import { parseVehicleImportResponse } from './parseVehicleImportResponse';
+import { buildVehiclePairXml } from '../api/buildVehiclePairXml';
+import { parseVehicleImportResponse } from '../api/parseVehicleImportResponse';
 import type { VehicleEditFormValue } from '../VehicleEditForm';
 
 interface SaveResult {
